% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_covid.R
\encoding{UTF-8}
\name{pnad_covid_microdata}
\alias{pnad_covid_microdata}
\title{Loading and compilating PNAD COVID19 microdata}
\usage{
pnad_covid_microdata(dir_data, lang = "eng", ...)
}
\arguments{
\item{dir_data}{Directory in which original .txt are located}

\item{lang}{In which language will the variables be labelled. Default, \code{lang = "eng"} is
in English. For Brazilian Portuguese, set \code{lang = "pt_br"}.}

\item{...}{vectors with survey dates as in  \code{c('MM', 'YYYY')}}
}
\value{
Dataframe list, with each entry corresponding to a date in \code{...}
}
\description{
Extracts and builds databases from the PNAD COVID19 survey in R format (.RData) from
the original microdata - for information on how to get the original data files, check IBGE's website -
www.ibge.gov.br .
Since the survey is still being published by IBGE,
this function is subject to frequent updates.
The function generates one database per month of the survey. If it is the case, use
the \code{\link{dplyr::bind_rows}} command to stack all data.
}
\examples{
pnad_covid_microdata('./Desktop', c(5, 2020))

}
